checking for existing user before configuring user

+++ build_scripts/start.sh
@@ -6,32 +6,39 @@ if [ "$3" != "" ]; then
     export USER_KEY=$3
 fi
 
-echo "configuring user: $USER_NAME ..."
+user_exists=$(id -u $USER_NAME > /dev/null 2>&1; echo $?) 
+if [[ "$user_exists" == 0 ]]; then
+    echo "user exists: $USER_NAME"
+else
+    echo "configuring user: $USER_NAME ..."
 
-sudo adduser --disabled-password --gecos '' --uid $USER_ID $USER_NAME > /dev/null 2>&1 
-sudo adduser $USER_NAME sudo > /dev/null 2>&1 
+    sudo adduser --disabled-password --gecos '' --uid $USER_ID $USER_NAME > /dev/null 2>&1 
+    sudo adduser $USER_NAME sudo > /dev/null 2>&1 
 
-sudo su $USER_NAME -c "mkdir \$HOME/.ssh"
-if [ "$USER_KEY" != "" ]; then
-    sudo su $USER_NAME -c "echo $USER_KEY > \$HOME/.ssh/authorized_keys"
-    sudo su $USER_NAME -c "chmod 600 \$HOME/.ssh/authorized_keys"
+    sudo su $USER_NAME -c "mkdir \$HOME/.ssh"
+    if [ "$USER_KEY" != "" ]; then
+        sudo su $USER_NAME -c "echo $USER_KEY > \$HOME/.ssh/authorized_keys"
+        sudo su $USER_NAME -c "chmod 600 \$HOME/.ssh/authorized_keys"
+    fi
+
+    WORKAREA=/home/builder/workarea/
+    cd $WORKAREA
+
+    sudo su $USER_NAME -c "cp tmux.conf ~/.tmux.conf"
+    sudo su $USER_NAME -c "cp vimrc ~/.vimrc"
+    sudo su $USER_NAME -c "cp myVimrc ~"
+    sudo su $USER_NAME -c "cp myBashrc ~"
+    sudo su $USER_NAME -c "echo '. ~/myBashrc' >> ~/.bashrc"
+
+    for s in .vim ; do
+        if [ -e $HOME/$s ]; then
+            cd $HOME/$s
+            sudo su $USER_NAME -c "find . -depth -print | cpio -pdvum ~/$s" > /dev/null 2>&1
+        fi
+    done
+
+    sudo su $USER_NAME -c "$WORKAREA/personalize.sh"
 fi
-
-WORKAREA=/home/builder/workarea/
-cd $WORKAREA
-
-sudo su $USER_NAME -c "cp tmux.conf ~/.tmux.conf"
-sudo su $USER_NAME -c "cp vimrc ~/.vimrc"
-sudo su $USER_NAME -c "cp myVimrc ~"
-sudo su $USER_NAME -c "cp myBashrc ~"
-sudo su $USER_NAME -c "echo '. ~/myBashrc' >> ~/.bashrc"
-
-for s in .vim .haste; do
-    cd $HOME/$s
-    sudo su $USER_NAME -c "find . -depth -print | cpio -pdvum ~/$s" > /dev/null 2>&1
-done
-
-sudo su $USER_NAME -c "$WORKAREA/personalize.sh"
 
 echo "sshd started"
 if [ "$USER_KEY" != "" ]; then
